Struct migration_engine_tests::multi_engine_test_api::TestApi

pub struct TestApi { /* fields omitted */ }

The multi-engine test API.

Implementations

impl TestApi

pub fn new(args: TestApiArgs) -> Self

Initializer, called by the test macros.

pub fn block_on<O, F: Future<Output = O>>(&self, f: F) -> O

Block on a future

pub fn query_raw(&self, sql: &str, params: &[Value<'_>]) -> Result<ResultSet>

Equivalent to quaint’s query_raw()

pub fn raw_cmd(&self, sql: &str)

Send a SQL command to the database, and expect it to succeed.

pub fn connection_string(&self) -> &str

The connection string for the database associated with the test.

pub fn connection_info(&self) -> ConnectionInfo

The ConnectionInfo based on the connection string

pub fn create_migrations_directory(&self) -> TempDir

Create a temporary directory to serve as a test migrations directory.

pub fn datasource_block(&self) -> DatasourceBlock<'_>

Render a valid datasource block, including database URL.

pub fn is_mssql(&self) -> bool

Returns true only when testing on MSSQL.

pub fn is_mysql(&self) -> bool

Returns true only when testing on MySQL.

pub fn is_mysql_mariadb(&self) -> bool

Returns true only when testing on MariaDB.

pub fn is_mysql_5_6(&self) -> bool

Returns true only when testing on MySQL 5.6.

pub fn is_mysql_8(&self) -> bool

Returns true only when testing on MySQL 8.

pub fn is_postgres(&self) -> bool

Returns true only when testing on postgres.

pub fn is_cockroach(&self) -> bool

Returns true only when testing on cockroach.

pub fn is_sqlite(&self) -> bool

Returns true only when testing on sqlite.

pub fn is_vitess(&self) -> bool

Returns true only when testing on vitess.

pub fn lower_cases_table_names(&self) -> bool

Returns whether the database automatically lower-cases table names.

pub fn new_engine(&self) -> EngineTestApi<'_>

Instantiate a new migration engine for the current database.

pub fn new_engine_with_connection_strings(
&self,
connection_string: String,
shadow_db_connection_string: Option<String>
) -> EngineTestApi<'_>

Instantiate a new migration with the provided connection string.

pub fn test_fn_name(&self) -> &str

The name of the test function, as a string.

pub fn datamodel_with_provider(&self, schema: &str) -> String

Render a datamodel including provider and generator block.

pub fn write_datasource_block(&self, out: &mut dyn Write)

Render a valid datasource block, including database URL.

pub fn preview_features(&self) -> BitFlags<PreviewFeature>

Currently enabled preview features.