Struct migration_engine_tests::test_api::TestApi

pub struct TestApi {
    pub connector: SqlMigrationConnector,
    // some fields omitted
}

Fields
connector: SqlMigrationConnector
Implementations

impl TestApi

pub fn new(args: TestApiArgs) -> Self

Initializer, called by the test macros.

pub fn args(&self) -> &TestApiArgs

pub fn apply_migrations<'a>(
&'a self,
migrations_directory: &'a TempDir
) -> ApplyMigrations<'a>

Plan an applyMigrations command

pub fn connection_string(&self) -> &str

pub fn connection_info(&self) -> ConnectionInfo

pub fn db_execute(&self, params: &DbExecuteParams) -> ConnectorResult<()>

pub fn ensure_connection_validity(&self) -> ConnectorResult<()>

pub fn schema_name(&self) -> String

pub fn create_migration<'a>(
&'a self,
name: &'a str,
schema: &'a str,
migrations_directory: &'a TempDir
) -> CreateMigration<'a>

Plan a createMigration command

pub fn create_migrations_directory(&self) -> TempDir

Create a temporary directory to serve as a test migrations directory.

pub fn dev_diagnostic<'a>(
&'a self,
migrations_directory: &'a TempDir
) -> DevDiagnostic<'a>

Builder and assertions to call the devDiagnostic command.

pub fn diagnose_migration_history<'a>(
&'a self,
migrations_directory: &'a TempDir
) -> DiagnoseMigrationHistory<'a>

pub fn diff(&self, params: &DiffParams) -> ConnectorResult<DiffResult>

pub fn dump_table(&self, table_name: &str) -> ResultSet

pub fn evaluate_data_loss<'a>(
&'a self,
migrations_directory: &'a TempDir,
schema: String
) -> EvaluateDataLoss<'a>

pub fn is_mssql(&self) -> bool

Returns true only when testing on MSSQL.

pub fn is_mariadb(&self) -> bool

Returns true only when testing on MariaDB.

pub fn is_mysql(&self) -> bool

Returns true only when testing on MySQL.

pub fn is_mysql_mariadb(&self) -> bool

Returns true only when testing on MariaDB.

pub fn is_mysql_5_6(&self) -> bool

Returns true only when testing on MySQL 5.6.

pub fn is_mysql_8(&self) -> bool

Returns true only when testing on MySQL 8.

pub fn is_postgres(&self) -> bool

Returns true only when testing on postgres.

pub fn is_cockroach(&self) -> bool

Returns true only when testing on cockroach.

pub fn is_sqlite(&self) -> bool

Returns true only when testing on sqlite.

pub fn is_vitess(&self) -> bool

Returns true only when testing on vitess.

pub fn insert<'a>(&'a self, table_name: &'a str) -> SingleRowInsert<'a>

Insert test values

pub fn list_migration_directories<'a>(
&'a self,
migrations_directory: &'a TempDir
) -> ListMigrationDirectories<'a>

pub fn lower_cases_table_names(&self) -> bool

pub fn mark_migration_applied<'a>(
&'a self,
migration_name: impl Into<String>,
migrations_directory: &'a TempDir
) -> MarkMigrationApplied<'a>

pub fn mark_migration_rolled_back(
&self,
migration_name: impl Into<String>
) -> MarkMigrationRolledBack<'_>

pub fn migration_persistence<'a>(&'a self) -> &(dyn MigrationPersistence + 'a)

pub fn assert_schema(&self) -> SchemaAssertion

Assert facts about the database schema

pub fn block_on<O, F: Future<Output = O>>(&self, f: F) -> O

Block on a future.

pub fn datasource_block(&self) -> DatasourceBlock<'_>

Render a valid datasource block, including database URL.

pub fn datasource_block_with<'a>(
&'a self,
params: &'a [(&'a str, &'a str)]
) -> DatasourceBlock<'a>

pub fn connector_diff(&self, from: DiffTarget<'_>, to: DiffTarget<'_>) -> String

Generate a migration script using MigrationConnector::diff().

pub fn normalize_identifier<'a>(&self, identifier: &'a str) -> Cow<'a, str>

pub fn query(&self, q: Query<'_>) -> ResultSet

Like quaint::Queryable::query()

pub fn query_raw(&self, q: &str, params: &[Value<'static>]) -> ResultSet

Like quaint::Queryable::query_raw()

pub fn raw_cmd(&self, sql: &str)

Send a SQL command to the database, and expect it to succeed.

pub fn render_table_name<'a>(&'a self, table_name: &'a str) -> Table<'a>

Render a table name with the required prefixing for use with quaint query building.

pub fn reset(&self) -> Reset<'_>

Plan a reset command

pub fn schema_push_w_datasource(&self, dm: impl Into<String>) -> SchemaPush<'_>

Plan a schemaPush command adding the datasource

pub fn schema_push(&self, dm: impl Into<String>) -> SchemaPush<'_>

Plan a schemaPush command

pub fn tags(&self) -> BitFlags<Tags>

pub fn write_datasource_block(&self, out: &mut dyn Write)

Render a valid datasource block, including database URL.

pub fn datamodel_with_provider(&self, schema: &str) -> String