Module migration_engine_tests::test_api

Macros

expect

Creates an instance of Expect from string literal:

Structs

BitFlags

Represents a set of flags of some type T. T must have the #[bitflags] attribute applied.

DbExecuteParams

The type of params accepted by dbExecute.

DiffParams

The type of params for the diff method.

DiffResult

The result type for the diff method.

SchemaAssertion

SchemaContainer

An object with a schema field.

SingleRowInsert

TestApi

TestConnectorHost

UrlContainer

An object with a url field.

Enums

Capabilities

Test-relevant connector capabilities.

DbExecuteDatasourceType

The location of the live database to connect to.

Tags

Test-relevant connector tags.

Traits

MigrationsAssertions

ResultSetExt

Functions

tok

Attribute Macros

test_connector